moved db.js to private folder to keep things cleaner.
