+++ index.js
@@ -1,6 +1,6 @@
 //Import modules
 var fs = require('fs');
-var db = require('./db.js');
+var db = require('./private/db.js');
 var https = require('https');
 var privateKey  = fs.readFileSync('private/certificates/server.key', 'utf8');
 var certificate = fs.readFileSync('private/certificates/server.crt', 'utf8');
